ark-core/duckdb/requirements.txt
duckdb==1.1.3
nats-py==2.9.0
pandas==2.2.3
pyarrow==17.0.0
